feat(oauth.route.ts): OAuth 라우트 및 스키마 업데이트
